feat(board): enable logging feature and fix related compilation issues
- Add direct `log` dependency to `board` crate to resolve unresolved module errors
- Configure `logging` feature properly in both `board` and `logrs` crates
- Update `.pre-commit-config.yaml` to exclude additional paths from secret scanning
- Introduce new dependencies for file upload functionality (e.g., `multer`, `mime_guess`)
- Refactor logging initialization to use `logrs::init_logger`
- Move test files from `src/tests` to `tests` directory for better organization
- Replace `tracing` macros with `logrs` re-exported logging macros
- Fix integration tests by updating HTTP method usage and router nesting
- Update documentation with detailed logging feature fix record

fix(sqlx): ensure compile-time query validation passes

- Create necessary database tables via migrations before compilation
- Set `DATABASE_URL` environment variable during build and SQLx preparation
- Generate `.sqlx` query data files using `cargo sqlx prepare --workspace`

chore(deps): update Cargo.lock with new and modified dependencies

- Add `log`, `mime`, `multer`, `sha2`, `tempfile`, `tokio-util`, `uuid`, and others
- Remove unused dependencies like `tower-http`, `tracing`, `tracing-subscriber` from root
- Adjust features and dependencies in `logrs` and `board` crates accordingly
