+++ Cargo.toml
@@ -2,8 +2,8 @@
 name = "elizabeth"
 version = "0.3.0"
 edition = "2024"
-authors=["unic <yuniqueunic@gmail.com>"]
-license="BSD-3"
+authors = ["unic <yuniqueunic@gmail.com>"]
+license = "BSD-3"
 
 [workspace]
 members = ["crates/*"]
@@ -11,5 +11,5 @@ resolver = "3"
 
 
 [workspace.dependencies]
-logrs = { path = "crates/logrs" }
+logrs = { path = "crates/logrs", features = ["logging"] }
 configrs = { path = "crates/configrs" }
