+++ crates/board/Cargo.toml
@@ -6,16 +6,18 @@ edition.workspace = true
 license.workspace = true
 
 [features]
-default = []
-full = ["completions"]
+default = ["logging"]
+full = ["completions", "logging"]
 dynamic = []
 # Optional features
 completions = ["dep:clap_complete"]
+logging = ["logrs/logging", "dep:log"]
 disable-log = ["logrs/disable-log"]
 structured-log = ["logrs/structured-log"]
 
 [dependencies]
-logrs = { workspace = true }
+log = { version = "0.4", optional = true }
+logrs = { workspace = true, features = ["logging"] }
 configrs = { workspace = true }
 
 shadow-rs = { version = "1", features = ["no_std"] }
@@ -25,9 +27,7 @@ clap_complete = { version = "4", optional = true }
 
 tokio = { version = "1.47.1", features = ["full"] }
 axum = "0.8.6"
-tower = "0.5"
-tower-http = "0.6"
-http = "1.1"
+axum_responses = "0.4.4"
 
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
@@ -35,21 +35,36 @@ serde_json = "1.0.145"
 anyhow = "1.0.100"
 thiserror = "2.0.17"
 
-
 utoipa = "5.4.0"
 utoipa-scalar = { version = "0.3.0", features = ["axum"] }
-
-
-tracing = "0.1.41"
-tracing-subscriber = "0.3.20"
-axum_responses = "0.4.4"
 utoipa-axum = "0.2.0"
+
 chrono = { version = "0.4.42", features = ["serde"] }
 
 
-sqlx = { version = "0.8", features = ["runtime-tokio", "tls-rustls-aws-lc-rs", "sqlite", "migrate", "chrono"] }
+sqlx = { version = "0.8", features = [
+    "runtime-tokio",
+    "tls-rustls-aws-lc-rs",
+    "sqlite",
+    "migrate",
+    "chrono",
+] }
 async-trait = "0.1"
 
+# JWT authentication
+uuid = { version = "1", features = ["v4", "serde"] }
+
+# File upload dependencies
+multer = "3"
+tokio-util = { version = "0.7", features = ["io"] }
+mime = "0.3"
+mime_guess = "2.0"
+sha2 = "0.10"
+sanitize-filename = "0.6"
+tempfile = "3.0"
+
+[dev-dependencies]
+tower = "0.5"
 
 [build-dependencies]
 anyhow = { version = "1" }
